+++ src/chart/generator/marker.cpp
@@ -1,5 +1,6 @@
 #include "marker.h"
 
+#include <cmath>
 #include <cstdint>
 #include <optional>
 #include <utility>
@@ -31,10 +32,10 @@ Marker::Marker(const Options &options,
     bool needMarkerInfo) :
     cellInfo(data.cellInfo(index, needMarkerInfo)),
     mainId(data.getId(mainAxisList,
-        options.dimLabelIndex(-options.mainAxisType()),
+        options.dimLabelIndex(+options.mainAxisType()),
         index)),
     subId(data.getId(subAxisList,
-        options.dimLabelIndex(-options.subAxisType()),
+        options.dimLabelIndex(+options.subAxisType()),
         index)),
     sizeId(data.getId(
         options.getChannels().at(ChannelId::size).dimensions(),
@@ -76,7 +77,7 @@ Marker::Marker(const Options &options,
 	if (subAxisList != options.subAxis().dimensions())
 		subId.label =
 		    data.getId(options.subAxis().dimensions(),
-		            options.dimLabelIndex(-options.subAxisType()),
+		            options.dimLabelIndex(+options.subAxisType()),
 		            index)
 		        .label;
 
@@ -223,14 +224,14 @@ void Marker::fromRectangle(const Geom::Rect &rect)
 
 Math::Range<> Marker::getSizeBy(AxisId axisId) const
 {
-	return isHorizontal(+axisId) ? toRectangle().hSize()
-	                             : toRectangle().vSize();
+	return isHorizontal(orientation(axisId)) ? toRectangle().hSize()
+	                                         : toRectangle().vSize();
 }
 
 void Marker::setSizeBy(AxisId axisId, const Math::Range<> range)
 {
 	auto rect = toRectangle();
-	if (isHorizontal(+axisId))
+	if (isHorizontal(orientation(axisId)))
 		rect.setHSize(range);
 	else
 		rect.setVSize(range);