+++ src/chart/options/channelrange.cpp
@@ -1,26 +1,9 @@
 #include "channelrange.h"
 
-#include <string>
-
-#include "base/conv/parse.h"
-#include "base/conv/tostring.h"
 #include "base/math/range.h"
-#include "base/text/valueunit.h"
 
 namespace Vizzu::Gen
 {
-
-ChannelExtrema::ChannelExtrema(const std::string &str)
-{
-	const Text::ValueUnit vu(str);
-	value = vu.getValue();
-	unit = Conv::parse<ChannelExtremaType>(vu.getUnit());
-}
-
-ChannelExtrema::operator std::string() const
-{
-	return std::to_string(value) + std::string{Conv::toString(unit)};
-}
 
 Math::Range<> ChannelRange::getRange(
     const Math::Range<> &original) const